:bug: (rabbitmq connection and code structures) not connect mq connection every api calls, and seperate files code
